// ClientApp/src/Entity/Enums/PermisionEnum.ts
export enum PermissionEnum {
    PermissionsAdsPermissionsCreate = "Permissions.AdsPermissions.Create",
    PermissionsAdsPermissionsRead = "Permissions.AdsPermissions.Read",
    PermissionsAdsPermissionsUpdate = "Permissions.AdsPermissions.Update",
    PermissionsAdsPermissionsDelete = "Permissions.AdsPermissions.Delete",
    PermissionsAdsPermissionsPublish = "Permissions.AdsPermissions.Publish",

    PermissionsCategoryPermissionsCreate = "Permissions.CategoryPermissions.Create",
    PermissionsCategoryPermissionsRead = "Permissions.CategoryPermissions.Read",
    PermissionsCategoryPermissionsUpdate = "Permissions.CategoryPermissions.Update",
    PermissionsCategoryPermissionsDelete = "Permissions.CategoryPermissions.Delete",

    PermissionsCrawlerPermissionsCreate = "Permissions.CrawlerPermissions.Create",
    PermissionsCrawlerPermissionsRead = "Permissions.CrawlerPermissions.Read",
    PermissionsCrawlerPermissionsUpdate = "Permissions.CrawlerPermissions.Update",
    PermissionsCrawlerPermissionsDelete = "Permissions.CrawlerPermissions.Delete",

    PermissionsFilePermissionsUpload = "Permissions.FilePermissions.Upload",
    PermissionsFilePermissionsRead = "Permissions.FilePermissions.Read",
    PermissionsFilePermissionsDelete = "Permissions.FilePermissions.Delete",

    PermissionsPostPermissionsCreate = "Permissions.PostPermissions.Create",
    PermissionsPostPermissionsRead = "Permissions.PostPermissions.Read",
    PermissionsPostPermissionsUpdate = "Permissions.PostPermissions.Update",
    PermissionsPostPermissionsDelete = "Permissions.PostPermissions.Delete",
    PermissionsPostPermissionsApprove = "Permissions.PostPermissions.Approve",

    PermissionsRolePermissionsCreate = "Permissions.RolePermissions.Create",
    PermissionsRolePermissionsRead = "Permissions.RolePermissions.Read",
    PermissionsRolePermissionsUpdate = "Permissions.RolePermissions.Update",
    PermissionsRolePermissionsDelete = "Permissions.RolePermissions.Delete",

    PermissionsTopicPermissionsCreate = "Permissions.TopicPermissions.Create",
    PermissionsTopicPermissionsRead = "Permissions.TopicPermissions.Read",
    PermissionsTopicPermissionsUpdate = "Permissions.TopicPermissions.Update",
    PermissionsTopicPermissionsDelete = "Permissions.TopicPermissions.Delete",

    PermissionsUserPermissionsCreate = "Permissions.UserPermissions.Create",
    PermissionsUserPermissionsRead = "Permissions.UserPermissions.Read",
    PermissionsUserPermissionsUpdate = "Permissions.UserPermissions.Update",
    PermissionsUserPermissionsDelete = "Permissions.UserPermissions.Delete",

    PermissionsVideoPermissionsCreate = "Permissions.VideoPermissions.Create",
    PermissionsVideoPermissionsRead = "Permissions.VideoPermissions.Read",
    PermissionsVideoPermissionsUpdate = "Permissions.VideoPermissions.Update",
    PermissionsVideoPermissionsDelete = "Permissions.VideoPermissions.Delete",
    PermissionsVideoPermissionsApprove = "Permissions.VideoPermissions.Approve",

    PermissionsVotePermissionsCreate = "Permissions.VotePermissions.Create",
    PermissionsVotePermissionsRead = "Permissions.VotePermissions.Read",
    PermissionsVotePermissionsUpdate = "Permissions.VotePermissions.Update",
    PermissionsVotePermissionsDelete = "Permissions.VotePermissions.Delete",
}